Enable add class
